+++ src/js/route.js
@@ -1,267 +1,367 @@
-import UrlBuilder from './UrlBuilder';
-import { stringify } from 'qs';
+import { parse, stringify } from 'qs';
 
+/**
+ * A Laravel route. This class represents one route and its configuration and metadata.
+ */
+class Route {
+    /**
+     * @param {String} name - Route name.
+     * @param {Object} definition - Route definition.
+     * @param {Object} config - Ziggy configuration.
+     */
+    constructor(name, definition, config) {
+        this.name = name;
+        this.definition = definition;
+        this.bindings = definition.bindings ?? {};
+        this.config = { absolute: true, ...config };
+    }
+
+    /**
+     * Get a 'template' of the complete URL for this route.
+     *
+     * @example
+     * https://{team}.ziggy.dev/user/{user}
+     *
+     * @return {String} Route template.
+     */
+    get template() {
+        // If  we're building just a path there's no origin, otherwise: if this route has a
+        // domain configured we construct the origin with that, if not we use the app URL
+        const origin = !this.config.absolute ? '' : this.definition.domain
+            ? `${this.config.baseProtocol}://${this.definition.domain}${this.config.basePort ? `:${this.config.basePort}` : ''}`
+            : this.config.baseUrl;
+
+        return `${origin}/${this.definition.uri}`;
+    }
+
+    /**
+     * Get an array of objects representing the parameters that this route accepts.
+     *
+     * @example
+     * [{ name: 'team', required: true }, { name: 'user', required: false }]
+     *
+     * @return {Array} Parameter segments.
+     */
+    get parameterSegments() {
+        return this.template.match(/{[^}?]+\??}/g)?.map((segment) => ({
+            name: segment.replace(/{|\??}/g, ''),
+            required: !/\?}$/.test(segment),
+        })) ?? [];
+    }
+
+    /**
+     * Get whether this route's template matches the given URL.
+     *
+     * @param {String} url - URL to check.
+     * @return {Boolean} Whether this route matches.
+     */
+    matchesUrl(url) {
+        if (!this.definition.methods.includes('GET')) return false;
+
+        // Transform the route's template into a regex that will match a hydrated URL,
+        // by replacing its parameter segments with matchers for parameter values
+        const pattern = this.template
+            .replace(/\/{[^}?]*\?}/g, '(\/[^/?]+)?')
+            .replace(/{[^}]+}/g, '[^/?]+')
+            .replace(/^\w+:\/\//, '');
+
+        return new RegExp(`^${pattern}$`).test(url.replace(/\/+$/, '').split('?').shift());
+    }
+
+    /**
+     * Hydrate and return a complete URL for this route with the given parameters.
+     *
+     * @param {Object} params
+     * @return {String}
+     */
+    compile(params) {
+        if (!this.parameterSegments.length) return this.template.replace(/\/+$/, '');
+
+        return this.template.replace(/{([^}?]+)\??}/g, (_, segment) => {
+            // If the parameter is missing but is not optional, throw an error
+            if ([null, undefined].includes(params[segment]) && this.parameterSegments.find(({ name }) => name === segment).required) {
+                throw new Error(`Ziggy error: '${segment}' parameter is required for route '${this.name}'.`)
+            }
+
+            return encodeURIComponent(params[segment] ?? '');
+        }).replace(/\/+$/, '');
+    }
+}
+
+/**
+ * A collection of Laravel routes. This class constitutes Ziggy's main API.
+ */
 class Router extends String {
-    constructor(name, params, absolute, customZiggy = null) {
+    /**
+     * @param {String} name - Route name.
+     * @param {(String|Number|Array|Object)} params - Route parameters.
+     * @param {Boolean} absolute - Whether to include the URL origin.
+     * @param {Object} config - Ziggy configuration.
+     */
+    constructor(name, params, absolute = true, config) {
         super();
 
-        this.name = name;
-        this.absolute = absolute;
-        this.ziggy = customZiggy ? customZiggy : Ziggy;
-        this.urlBuilder = this.name ? new UrlBuilder(name, absolute, this.ziggy) : null;
-        this.template = this.urlBuilder ? this.urlBuilder.construct() : '';
-        this.urlParams = this.normalizeParams(params);
-        this.queryParams = {};
-        this.hydrated = '';
-    }
+        this._config = config ?? Ziggy ?? globalThis?.Ziggy;
 
-    normalizeParams(params) {
-        if (typeof params === 'undefined') return {};
-
-        // If you passed in a string or integer, wrap it in an array
-        params = typeof params !== 'object' ? [params] : params;
-
-        this.numericParamIndices = Array.isArray(params);
-        return Object.assign({}, params);
-    }
-
-    with(params) {
-        this.urlParams = this.normalizeParams(params);
-        return this;
-    }
-
-    withQuery(params) {
-        Object.assign(this.queryParams, params);
-        return this;
-    }
-
-    hydrateUrl() {
-        if (this.hydrated) return this.hydrated;
-
-        let hydrated = this.template.replace(
-            /{([^}]+)}/gi,
-            (tag, i) => {
-                let keyName = this.trimParam(tag),
-                    defaultParameter,
-                    tagValue;
-
-                if (this.ziggy.defaultParameters.hasOwnProperty(keyName)) {
-                    defaultParameter = this.ziggy.defaultParameters[keyName];
-                }
-
-                // If a default parameter exists, and a value wasn't
-                // provided for it manually, use the default value
-                if (defaultParameter && !this.urlParams[keyName]) {
-                    delete this.urlParams[keyName];
-                    return defaultParameter;
-                }
-
-                // We were passed an array, shift the value off the
-                // object and return that value to the route
-                if (this.numericParamIndices) {
-                    this.urlParams = Object.values(this.urlParams);
-
-                    tagValue = this.urlParams.shift();
-                } else {
-                    tagValue = this.urlParams[keyName];
-                    delete this.urlParams[keyName];
-                }
-
-                // The block above is what requires us to assign tagValue below
-                // instead of returning - if multiple *objects* are passed as
-                // params, numericParamIndices will be true and each object will
-                // be assigned above, which means !tagValue will evaluate to
-                // false, skipping the block below.
-
-                // If a value wasn't provided for this named parameter explicitly,
-                // but the object that was passed contains an ID, that object
-                // was probably a model, so we use the ID.
-
-                let bindingKey = this.ziggy.namedRoutes[this.name]?.bindings?.[keyName];
-
-                if (bindingKey && !this.urlParams[keyName] && this.urlParams[bindingKey]) {
-                    tagValue = this.urlParams[bindingKey];
-                    delete this.urlParams[bindingKey];
-                } else if (!tagValue && !this.urlParams[keyName] && this.urlParams['id']) {
-                    tagValue = this.urlParams['id']
-                    delete this.urlParams['id'];
-                }
-
-                // The value is null or undefined; is this param
-                // optional or not
-                if (tagValue == null) {
-                    if (tag.indexOf('?') === -1) {
-                        throw new Error(
-                            "Ziggy Error: '" +
-                                keyName +
-                                "' key is required for route '" +
-                                this.name +
-                                "'"
-                        );
-                    } else {
-                        return '';
-                    }
-                }
-
-                // If an object was passed and has an id, return it
-                if (tagValue.id) {
-                    return encodeURIComponent(tagValue.id);
-                } else if (tagValue[bindingKey]) {
-                    return encodeURIComponent(tagValue[bindingKey])
-                }
-
-                return encodeURIComponent(tagValue);
+        if (name) {
+            if (!this._config.namedRoutes[name]) {
+                throw new Error(`Ziggy error: route '${name}' is not in the route list.`);
             }
-        );
 
-        if (this.urlBuilder != null && this.urlBuilder.path !== '') {
-          hydrated = hydrated.replace(/\/+$/, '');
+            this._route = new Route(name, this._config.namedRoutes[name], { ...this._config, absolute });
+            this._params = this._parse(params);
         }
-
-        this.hydrated = hydrated;
-
-        return this.hydrated;
     }
 
-    matchUrl() {
-        let windowUrl =
-            window.location.hostname +
-            (window.location.port ? ':' + window.location.port : '') +
-            window.location.pathname;
+    /**
+     * Get the compiled URL string for the current route and parameters.
+     *
+     * @example
+     * // with 'posts.show' route 'posts/{post}'
+     * route('posts.show', 1).url(); // 'https://ziggy.dev/posts/1'
+     *
+     * @return {String}
+     */
+    url() {
+        // Get parameters that don't correspond to any route segments to append them to the query
+        const unhandled = Object.keys(this._params)
+            .filter((key) => !this._route.parameterSegments.some(({ name }) => name === key))
+            .reduce((result, current) => ({ ...result, [current]: this._params[current] }), {});
 
-        // Strip out optional parameters
-        let optionalTemplate = this.template
-            .replace(/(\/\{[^\}]*\?\})/g, '/')
-            .replace(/(\{[^\}]*\})/gi, '[^/?]+')
-            .replace(/\/?$/, '')
-            .split('://')[1];
-
-        let searchTemplate = this.template
-            .replace(/(\{[^\}]*\})/gi, '[^/?]+')
-            .split('://')[1];
-        let urlWithTrailingSlash = windowUrl.replace(/\/?$/, '/');
-
-        const regularSearch = new RegExp('^' + searchTemplate + '/$').test(
-            urlWithTrailingSlash
-        );
-        const optionalSearch = new RegExp('^' + optionalTemplate + '/$').test(
-            urlWithTrailingSlash
-        );
-
-        return regularSearch || optionalSearch;
-    }
-
-    constructQuery() {
-        if (
-            Object.keys(this.queryParams).length === 0 &&
-            Object.keys(this.urlParams).length === 0
-        ) {
-            return '';
-        }
-
-        let remainingParams = Object.assign(this.urlParams, this.queryParams);
-
-        return stringify(remainingParams, {
+        return this._route.compile(this._params) + stringify({ ...unhandled, ...this._queryParams }, {
+            addQueryPrefix: true,
+            arrayFormat: 'indices',
             encodeValuesOnly: true,
             skipNulls: true,
-            addQueryPrefix: true,
-            arrayFormat: 'indices'
         });
     }
 
-    current(name = null) {
-        let routeNames = Object.keys(this.ziggy.namedRoutes);
+    /**
+     * Get the name of the route matching the current window URL, or, given a route name
+     * and parameters, check if the current window URL and parameters match that route.
+     *
+     * @example
+     * // at URL https://ziggy.dev/posts/4 with 'posts.show' route 'posts/{post}'
+     * route().current(); // 'posts.show'
+     * route().current('posts.index'); // false
+     * route().current('posts.show'); // true
+     * route().current('posts.show', { post: 1 }); // false
+     * route().current('posts.show', { post: 4 }); // true
+     *
+     * @param {String} name - Route name to check.
+     * @param {(String|Number|Array|Object)} params - Route parameters.
+     * @return {(Boolean|String)}
+     */
+    current(name, params) {
+        const url = window.location.host + window.location.pathname;
 
-        let currentRoute = routeNames.filter(name => {
-            if (this.ziggy.namedRoutes[name].methods.indexOf('GET') === -1) {
-                return false;
-            }
+        // Find the first route that matches the current URL
+        const [current, route] = Object.entries(this._config.namedRoutes).find(
+            ([_, route]) => new Route(name, route, this._config).matchesUrl(url)
+        );
 
-            return new Router(
-                name,
-                undefined,
-                undefined,
-                this.ziggy
-            ).matchUrl();
-        })[0];
+        // If a name wasn't passed, return the name of the current route
+        if (!name) return current;
 
-        if (name) {
-            const pattern = new RegExp(
-                '^' + name.replace('.', '\\.').replace('*', '.*') + '$',
-                'i'
-            );
-            return pattern.test(currentRoute);
+        // Test the passed name against the current route, matching some
+        // basic wildcards, e.g. passing `events.*` matches `events.show`
+        const match = new RegExp(`^${name.replace('.', '\\.').replace('*', '.*')}$`).test(current);
+
+        if (!params) return match;
+
+        params = this._parse(params, new Route(current, route, this._config));
+
+        // Check that all passed parameters match their values in the current window URL
+        return Object.entries(this._dehydrate(route))
+            .filter(([key]) => params.hasOwnProperty(key))
+            // Use weak equality because all values in the current window URL will be strings
+            .every(([key, value]) => params[key] == value);
+    }
+
+    /**
+     * Get all parameter values from the current window URL.
+     *
+     * @example
+     * // at URL https://tighten.ziggy.dev/posts/4?lang=en with 'posts.show' route 'posts/{post}' and domain '{team}.ziggy.dev'
+     * route().params; // { team: 'tighten', post: 4, lang: 'en' }
+     *
+     * @return {Object}
+     */
+    get params() {
+        return this._dehydrate(this._config.namedRoutes[this.current()]);
+    }
+
+    /**
+     * Check whether the given route exists.
+     *
+     * @param {String} name
+     * @return {Boolean}
+     */
+    has(name) {
+        return Object.keys(this._config.namedRoutes).includes(name);
+    }
+
+    /**
+     * Add query parameters to be appended to the compiled URL.
+     *
+     * @param {Object} params
+     * @return {this}
+     */
+    withQuery(params) {
+        this._queryParams = params;
+        return this;
+    }
+
+    /**
+     * Parse Laravel-style route parameters of any type into a normalized object.
+     *
+     * @example
+     * // with route parameter names 'event' and 'venue'
+     * _parse(1); // { event: 1 }
+     * _parse({ event: 2, venue: 3 }); // { event: 2, venue: 3 }
+     * _parse(['Taylor', 'Matt']); // { event: 'Taylor', venue: 'Matt' }
+     * _parse([4, { uuid: 56789, name: 'Grand Canyon' }]); // { event: 4, venue: 56789 }
+     *
+     * @param {(String|Number|Array|Object)} params - Route parameters.
+     * @param {Route} route - Route instance.
+     * @return {Object} Normalized complete route parameters.
+     */
+    _parse(params = {}, route = this._route) {
+        // If `params` is a string or integer, wrap it in an array
+        params = ['string', 'number'].includes(typeof params) ? [params] : params;
+
+        // Separate segments with and without defaults, and fill in the default values
+        const segments = route.parameterSegments.filter(({ name }) => !this._config.defaultParameters[name]);
+
+        if (Array.isArray(params)) {
+            // If the parameters are an array they have to be in order, so we can transform them into
+            // an object by keying them with the template segment names in the order they appear
+            params = params.reduce((result, current, i) => ({ ...result, [segments[i].name]: current }), {});
+        } else if (
+            segments.length === 1
+            && !params[segments[0].name]
+            && (params.hasOwnProperty(Object.values(route.bindings)[0]) || params.hasOwnProperty('id'))
+        ) {
+            // If there is only one template segment and `params` is an object, that object is
+            // ambiguous—it could contain the parameter key and value, or it could be an object
+            // representing just the value (e.g. a model); we can inspect it to find out, and
+            // if it's just the parameter value, we can wrap it in an object with its key
+            params = { [segments[0].name]: params };
         }
 
-        return currentRoute;
+        return {
+            ...this._defaults(route),
+            ...this._substituteBindings(params, route.bindings),
+        };
     }
 
-    check(name) {
-        let routeNames = Object.keys(this.ziggy.namedRoutes);
-
-        return routeNames.includes(name);
+    /**
+     * Populate default parameters for the given route.
+     *
+     * @example
+     * // with default parameters { locale: 'en', country: 'US' } and 'posts.show' route '{locale}/posts/{post}'
+     * defaults(...); // { locale: 'en' }
+     *
+     * @param {Route} route
+     * @return {Object} Default route parameters.
+     */
+    _defaults(route) {
+        return route.parameterSegments.filter(({ name }) => this._config.defaultParameters[name])
+            .reduce((result, { name }, i) => ({ ...result, [name]: this._config.defaultParameters[name] }), {});
     }
 
-    extractParams(uri, template, delimiter) {
-        const uriParts = uri.split(delimiter);
-        const templateParts = template.split(delimiter);
+    /**
+     * Substitute Laravel route model bindings in the given parameters.
+     *
+     * @example
+     * _substituteBindings({ post: { id: 4, slug: 'hello-world', title: 'Hello, world!' } }, { post: 'slug' }); // { post: 'hello-world' }
+     *
+     * @param {Object} params - Route parameters.
+     * @param {Object} bindings - Route model bindings.
+     * @return {Object} Normalized route parameters.
+     */
+    _substituteBindings(params, bindings = {}) {
+        return Object.entries(params).reduce((result, [key, value]) => {
+            // If the value isn't an object there's nothing to substitute, so we return it as-is
+            if (!value || typeof value !== 'object' || Array.isArray(value)) {
+                return { ...result, [key]: value };
+            }
 
-        return templateParts.reduce(
-            (params, param, i) =>
-                param.indexOf('{') === 0 &&
-                param.indexOf('}') !== -1 &&
-                uriParts[i]
-                    ? Object.assign(params, {
-                          [this.trimParam(param)]: uriParts[i]
-                      })
-                    : params,
-            {}
-        );
+            if (!value.hasOwnProperty(bindings[key])) {
+                if (value.hasOwnProperty('id')) {
+                    // As a fallback, we still accept an 'id' key not explicitly registered as a binding
+                    bindings[key] = 'id';
+                } else {
+                    throw new Error(`Ziggy error: object passed as '${key}' parameter is missing route model binding key '${bindings[key]}'.`)
+                }
+            }
+
+            return { ...result, [key]: value[bindings[key]] };
+        }, {});
     }
 
-    get params() {
-        const namedRoute = this.ziggy.namedRoutes[this.current()];
-
+    /**
+     * Get all parameters and their values from the current window URL, based on the given route definition.
+     *
+     * @example
+     * // at URL https://tighten.ziggy.dev/events/8/venues/chicago?zoom=true
+     * _dehydrate({ domain: '{team}.ziggy.dev', uri: 'events/{event}/venues/{venue?}' }); // { team: 'tighten', event: 8, venue: 'chicago', zoom: true }
+     *
+     * @param {Object} route - Route definition.
+     * @return {Object} Parameters.
+     */
+    _dehydrate(route) {
         let pathname = window.location.pathname
-            .replace(this.ziggy.baseUrl.split('://')[1].split('/')[1], '')
+            // If this Laravel app is in a subdirectory, trim the subdirectory from the path
+            .replace(this._config.baseUrl.replace(/^\w*:\/\/[^/]+/, ''), '')
             .replace(/^\/+/, '');
 
-        return Object.assign(
-            this.extractParams(
-                window.location.hostname,
-                namedRoute.domain || '',
-                '.'
-            ),
-            this.extractParams(
-                pathname,
-                namedRoute.uri,
-                '/'
-            )
-        );
-    }
+        // Given part of a valid 'hydrated' URL containing all its parameter values,
+        // a route template, and a delimiter, extract the parameters as an object
+        // E.g. dehydrate('events/{event}/{venue}', 'events/2/chicago', '/'); // { event: 2, venue: 'chicago' }
+        const dehydrate = (hydrated, template = '', delimiter) => {
+            const [values, segments] = [hydrated, template].map(s => s.split(delimiter));
 
-    parse() {
-        this.return = this.hydrateUrl() + this.constructQuery();
-    }
+            return segments.reduce((result, current, i) => {
+                // Only include template segments that are route parameters
+                // AND have a value present in the passed hydrated string
+                return /^{[^}?]+\??}$/.test(current) && values[i]
+                    ? { ...result, [current.replace(/^{|\??}$/g, '')]: values[i] }
+                    : result;
+            }, {});
+        }
 
-    url() {
-        this.parse();
-        return this.return;
+        return {
+            ...dehydrate(window.location.host, route.domain, '.'), // Domain parameters
+            ...dehydrate(pathname, route.uri, '/'), // Path parameters
+            ...parse(window.location.search?.replace(/^\?/, '')), // Query parameters
+        };
     }
 
     toString() {
         return this.url();
     }
 
-    trimParam(param) {
-        return param.replace(/{|}|\?/g, '');
-    }
-
     valueOf() {
         return this.url();
     }
+
+    /**
+     * @deprecated since v1.0, pass parameters as the second argument to `route()` instead.
+     */
+    with(params) {
+        this._params = this._parse(params);
+        return this;
+    }
+
+    /**
+     * @deprecated since v1.0, use `has()` instead.
+     */
+    check(name) {
+        return this.has(name);
+    }
 }
 
-export default function route(name, params, absolute, customZiggy) {
-    return new Router(name, params, absolute, customZiggy);
+export default function route(name, params, absolute, config) {
+    return new Router(name, params, absolute, config);
 }