+++ tests/Unit/ZiggyTest.php
@@ -391,7 +391,7 @@ class ZiggyTest extends TestCase
         $ziggy = new Ziggy;
 
         $expected = [
-            'baseUrl' => 'http://ziggy.dev/',
+            'baseUrl' => 'http://ziggy.dev',
             'baseProtocol' => 'http',
             'baseDomain' => 'ziggy.dev',
             'basePort' => null,
@@ -438,7 +438,7 @@ class ZiggyTest extends TestCase
         ]]);
 
         $expected = [
-            'baseUrl' => 'http://ziggy.dev/',
+            'baseUrl' => 'http://ziggy.dev',
             'baseProtocol' => 'http',
             'baseDomain' => 'ziggy.dev',
             'basePort' => null,
@@ -453,7 +453,7 @@ class ZiggyTest extends TestCase
 
         $this->addPostCommentsRouteWithBindings($expected['namedRoutes']);
 
-        $json = '{"baseUrl":"http:\/\/ziggy.dev\/","baseProtocol":"http","baseDomain":"ziggy.dev","basePort":null,"defaultParameters":[],"namedRoutes":{"postComments.index":{"uri":"posts\/{post}\/comments","methods":["GET","HEAD"]}}}';
+        $json = '{"baseUrl":"http:\/\/ziggy.dev","baseProtocol":"http","baseDomain":"ziggy.dev","basePort":null,"defaultParameters":[],"namedRoutes":{"postComments.index":{"uri":"posts\/{post}\/comments","methods":["GET","HEAD"]}}}';
 
         if ($this->laravelVersion(7)) {
             $json = str_replace(